# Assorted minor fixes and performance improvements