--- a/commit.go
+++ b/commit.go
@@ -147,10 +147,10 @@ func (repos *Repository) LookupCommit(oid *Oid) (*Commit, error) {
 		return nil, err
 	}
 	tree, err := parseTreeData(data)
-	tree.Oid = ci.TreeId()
 	if err != nil {
 		return nil, err
 	}
+	tree.Oid = ci.TreeId()
 	tree.repository = repos
 	ci.Tree = tree
 	return ci, nil