--- a/oid.go
+++ b/oid.go
@@ -12,17 +12,9 @@ type Oid struct {
 }
 
 // Create a new Oid from a Sha1 string of length 40.
+// In performance-sensitive paths, use NewOidFromByteString.
 func NewOidFromString(sha1 string) (*Oid, error) {
-	b, err := hex.DecodeString(sha1)
-	if err != nil {
-		return nil, err
-	}
-	o := new(Oid)
-	for i := 0; i < 20; i++ {
-		o.Bytes[i] = b[i]
-	}
-
-	return o, nil
+	return NewOidFromByteString([]byte(sha1))
 }
 
 // Create a new Oid from a 20 byte slice.
@@ -30,20 +22,22 @@ func NewOid(b []byte) (*Oid, error) {
 	if len(b) != 20 {
 		return nil, errors.New("Length must be 20")
 	}
-	o := new(Oid)
-	for i := 0; i < 20; i++ {
-		o.Bytes[i] = b[i]
-	}
-	return o, nil
+	var o Oid
+	copy(o.Bytes[:], b)
+	return &o, nil
 }
 
-// Create a new Oid from a 40 byte slice representing a string. This saves calling
-// string(data) every time we need a new Oid
+// Create a new Oid from a 40 byte hex-encoded slice.
 func NewOidFromByteString(b []byte) (*Oid, error) {
 	if len(b) != 40 {
-		return nil, errors.New(fmt.Sprintf("Length must be 40, but is %d", len(b)))
+		return nil, fmt.Errorf("bad hex-encoded sha1 length %d want 40", len(b))
 	}
-	return NewOidFromString(string(b))
+	var o Oid
+	_, err := hex.Decode(o.Bytes[:], b)
+	if err != nil {
+		return nil, err
+	}
+	return &o, nil
 }
 
 // Create a new Oid from a 20 byte array
@@ -53,21 +47,10 @@ func NewOidFromArray(a SHA1) *Oid {
 
 // Return string (hex) representation of the Oid
 func (o *Oid) String() string {
-	result := make([]byte, 0, 40)
-	hexvalues := []byte("0123456789abcdef")
-	for i := 0; i < 20; i++ {
-		result = append(result, hexvalues[o.Bytes[i]>>4])
-		result = append(result, hexvalues[o.Bytes[i]&0xf])
-	}
-	return string(result)
+	return hex.EncodeToString(o.Bytes[:])
 }
 
-// Return true if oid2 has the same sha1 as caller
+// Equal reports whether o and oid2 have the same sha1.
 func (o *Oid) Equal(oid2 *Oid) bool {
-	for i, v := range oid2.Bytes {
-		if o.Bytes[i] != v {
-			return false
-		}
-	}
-	return true
+	return o.Bytes == oid2.Bytes
 }