Performance impact of the entire PR:

```
name                            old time/op    new time/op    delta
LookupReference/HEAD-8            13.5µs ±15%    12.6µs ± 2%   -6.47%         (p=0.008 n=10+9)
LookupReference/master-8          6.47µs ± 3%    6.47µs ± 3%     ~             (p=0.931 n=9+9)
LookupReference/doesnotexist-8    17.7µs ± 1%    18.1µs ± 4%   +2.32%         (p=0.009 n=8+10)

name                            old alloc/op   new alloc/op   delta
LookupReference/HEAD-8            2.29kB ± 0%    2.16kB ± 0%   -5.59%        (p=0.000 n=10+10)
LookupReference/master-8          1.22kB ± 0%    1.10kB ± 0%  -10.46%        (p=0.000 n=10+10)
LookupReference/doesnotexist-8    9.02kB ± 0%    9.02kB ± 0%     ~     (all samples are equal)

name                            old allocs/op  new allocs/op  delta
LookupReference/HEAD-8              19.0 ± 0%      16.0 ± 0%  -15.79%        (p=0.000 n=10+10)
LookupReference/master-8            11.0 ± 0%       8.0 ± 0%  -27.27%        (p=0.000 n=10+10)
LookupReference/doesnotexist-8      17.0 ± 0%      17.0 ± 0%     ~     (all samples are equal)
```