@pgundlach 2016-12-30
Thanks again, and sorry for the delay (I was travelling).